requirements.txt
Django
django-uuidfield
pyOpenSSL==0.14
